move scripts out of source
