+++ .github/workflows/deploy.yml
@@ -26,10 +26,10 @@ jobs:
           folder: '.'
 
       - name: Setup npm
-        run: npm ci --omit dev && node source/scripts/initialize.js
+        run: npm ci --omit dev && node scripts/initialize.js
 
       - name: Update Commands
-        run: node source/scripts/updateWiki.js
+        run: node scripts/updateWiki.js
 
       - name: Push Changes to Repo
         run: |
