+++ package.json
@@ -1,9 +1,9 @@
 {
 	"scripts": {
 		"start": "node index.js production",
-		"start:dev": "node source/scripts/uploadCommandsToTestGuild.js && node index.js",
-		"clearCommands": "node source/scripts/clearCommands.js",
-		"initialize": "npm ci && node source/scripts/initialize"
+		"start:dev": "node scripts/uploadCommandsToTestGuild.js && node index.js",
+		"clearCommands": "node scripts/clearCommands.js",
+		"initialize": "npm ci && node scripts/initialize"
 	},
 	"dependencies": {
 		"discord.js": "^14.18.0",
