+++ scripts/initialize.js
